Reduce memory usage and add ability to select if use double or float numbers

+++ include/data/pos.h
@@ -1,18 +1,20 @@
+#include <memory>
+
 struct Pos 
 {
     private:
-        unsigned int m_i;
-        unsigned int m_j;
+        uint8_t m_i;
+        uint8_t m_j;
 
     public:
         Pos() = delete;
         Pos(const Pos& other) : m_i(other.m_i), m_j(other.m_j) {}
-        Pos(unsigned int i, unsigned int j) : m_i(i), m_j(j) {}
+        Pos(uint8_t i, uint8_t j) : m_i(i), m_j(j) {}
 
-        unsigned int get_i();
-        unsigned int get_j();
+        uint8_t get_i();
+        uint8_t get_j();
 
-        void set_i(unsigned int i);
-        void set_j(unsigned int j);
-        void set(unsigned int i, unsigned int j);
+        void set_i(uint8_t i);
+        void set_j(uint8_t j);
+        void set(uint8_t i, uint8_t j);
 };
